pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openpm-env"
version = "0.1.0"
description = "OpenPM: deterministic project management environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "openpm_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["openpm_env", "openpm_env.server", "openpm_env.tasks"]

[tool.pytest.ini_options]
testpaths = ["tests"]
